+++ CHANGELOG.md
@@ -11,11 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Test module for deprecation warnings
 - Option to control the switching point of `TwoPhaseStrategy` (former `Strategy`)
 - `SequentialStrategy` and `StreamingSequentialStrategy` classes
+- Telemetry env variable `BAYBE_TELEMETRY_VPN_CHECK` turning the initial connectivity check on/off 
+- Telemetry env variable `BAYBE_TELEMETRY_VPN_CHECK_TIMEOUT` for setting the connectivity check timeout
 
 ### Changed
 - Reorganized modules into subpackages
 - Serialization no longer relies on cattrs' global converter
 - Refined (un-)structuring logic
+- Telemetry env variable `BAYBE_TELEMETRY_HOST` renamed to `BAYBE_TELEMETRY_ENDPOINT`
+- Telemetry env variable `BAYBE_DEBUG_FAKE_USERHASH` renamed to `BAYBE_TELEMETRY_USERNAME`
+- Telemetry env variable `BAYBE_DEBUG_FAKE_HOSTHASH` renamed to `BAYBE_TELEMETRY_HOSTNAME`
 
 ### Fixed
 - Now supports Python 3.11